harmonisatie/requirements.txt
flask==3.0.2
openai==1.12.0
python-dotenv==1.0.1
PyPDF2==3.0.1
werkzeug==3.0.1 